integration: correct naming for tests which use a wrapper function
This adds the `TestDispatch` interface, which gives the caller the opportunity
to nominate a function to use to calculate the name. This is useful if the real
test function takes arguments other than `*testing.T, Sandbox` since a wrapper
(possibly anonymous) can then be used to set things up and produce the extra
arguments etc.

The existing `Test` type implements the new interface and a simple helper
function means existing code continues to work.

Signed-off-by: Ian Campbell <ijc@docker.com>