dockerfile: add testing in `gateway` mode to the existing tests of `builtin`
To achive this each test case now takes an additional `frontend` parameter
which provides an appropriate `solver` and `dfCmdArgs` which are the only
places so far that need to differ based on the way the frontend is run.

Uses the new `TestDispatcher` interface to allow this new option to be added.

Signed-off-by: Ian Campbell <ijc@docker.com>