containerimage exporter: add 'registry.ca' exporter option to specify an additional ca certificate
This allows us to specify for example the internal kubernetes ca certificate to be used to verify pushes against an internal registry deployment.
Without this option the only way to achieve this would be to manipulate buildkitd's base image (copy ca file + run update-ca-certificates), and even this doesn't work out of the box.