Handle windows-layer mounts somewhat like bind-mounts
windows-layer mounts (the only mount type supported by Windows) do not
currently use the `target` parameter to `mount.All`.

So instead, we treat them like bind-mounts and access the activated
layer in-place. Unlike the bind-mounts, we still need to Mount and
Unmount the layer.

Since we don't own the directory, we must not delete it after we unmount
it.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>