Fix COPY --from=stage when running with user-namespaces
Issue was reported in moby/moby issue 34645, and relates to moby/moby PR 38599,
which addressed this issue for the classic builder.

When copying files between stages, file ownership should be preserved;

    FROM busybox:latest as build
    RUN touch x && chown 1:1 x

    FROM busybox:latest
    RUN touch y && chown 1:1 y
    COPY --from=build x ./

However, when running with user namespaces enabled:

    # grep dockremap /etc/sub*id
    /etc/subgid:dockremap:500000:65536
    /etc/subuid:dockremap:500000:65536

A build would fail, because BuildKit is trying to map the container's UID/GID
to the host:

    ERROR: failed to copy file info: failed to chown /var/lib/docker/500000.500000/overlay2/mu5zpgelkig84eo9rlbqdhnn9/merged/x: Container ID 500001 cannot be mapped to a host ID

When looking at the code, I found that `(fb *Backend) Copy()` takes a `action`
argument (`pb.FileActionCopy`), one of its options is wether or not the owner
should be overridden:

    // optional owner override
    Owner *ChownOpt `protobuf:"bytes,3,opt,name=owner,proto3" json:"owner,omitempty"`)

That argument is passed on to `docopy()` (which is called from `(fb *Backend) Copy()`),
however inside `docopy()`, user-mapping is performed, irregardless if `Owner` was set
or not:

    ch, err := mapUserToChowner(u, idmap)
    if err != nil {
    	return err
    }

    opt := []copy.Opt{
    	func(ci *copy.CopyInfo) {
    		ci.Chown = ch

This patch makes that step optional, and only performs `mapUserToChowner` if
`action.Owner` is specified.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>