Add support for apparmor/selinux
Set's an apparmor profile on the OCI spec if one is configred on the
worker.
Adds selinux labels to containers (only added if selinux is enabled on
the system).

This assumes that the specified apparmor profile is already loaded on
the system and does not try to load it or even check if it is loaded.

SELinux support requires the `selinux` build tag to be added.
Likewise, `runc` would require both the `apparmor` and `selinux` build
tags.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit 72ef6c9b5e5f8578f7bf963fd681d47f6468678c)

Vendored go-selinux v1.3.1 (version used by docker/docker 19.03)
Had to adapt code due to refactors.

Signed-off-by: Tibor Vass <tibor@docker.com>