Update buildkit (#40)
* contenthash: include basename in content checksum for wildcards

While we generally ignore the basename in this layer, for wildcards
there in no other place to add the basename to the checksum as they
can not be resolved earlier. Before the basename that was in the
checksum was the wildcard itself, so if the wildcard remained same,
content remained same but the file where wildcard pointed to was
renamed, the cache was not invalidated.

Unfortunately, this change breaks cache for all copy commands that
use a wildcard.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>

* Fix missing mounts in execOp cache map

A bug in cloneExecOp prevented mounts from being included in the digest
computed for the execOp cache map. This could lead to an exec being
wrongly cached when a different mount was used for a subsequent
execution.

Repro case:
https://gist.github.com/aaronlehmann/cfeaefc028df8127fb85b9b5f9125f2d

In this example, pass2 should generate an empty diff because the /from
and /to mounts are the same busybox image. But before the fix, it uses
the cached result from pass1 (with different mounts) instead.

Signed-off-by: Aaron Lehmann <alehmann@netflix.com>

* Add regression test for run cache not considering mounts

Signed-off-by: Aaron Lehmann <alehmann@netflix.com>

* Add hack to preserve Dockerfile RUN cache compatibility after mount cache bugfix

Signed-off-by: Aaron Lehmann <alehmann@netflix.com>

* Fix potential crash on nil value

Signed-off-by: Pierre Fenoll <pierrefenoll@gmail.com>

* Dockerfile: update containerd to v1.5.0-rc.2 / v1.4.4

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>

* go.mod: containerd v1.5.0-rc.2

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>

* Remove potentially expired tokens.

Some registries (notably Quay) issue tokens that expire without providing an `expires in` value in the authorization payload. Therefore, if a token produces a 401, we should remove it and re-fetch.

Signed-off-by: Corey Larson <corey@earthly.dev>

* Add "ALL_PROXY" to list of default args / proxy env-vars

Relates to https://github.com/linuxkit/linuxkit/blob/a82fff637781877fc0d88c0f6e514fcd24699c46/docs/packages.md#proxies

> (..) the first four of these are the standard built-in build-arg options
> available for `docker build`
> (..) The last, `all_proxy`, is a standard var used for socks proxying. Since
> it is not built into `docker build`, if you want to use it, you will need to
> add the following line to the dockerfile:
>
>     ARG all_proxy

Given the we support all other commonly known proxy env-vars by default, it makes
sense to add this one as well.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

* Redact credentials from URLs before returning errors

this is to prevent errors such as

    failed to fetch remote https://user:password@github.com/user/private-repo-failure.git: exit status 128

from leaking the password; now it will be displayed like:

    failed to fetch remote https://user:xxxxx@non-existant-host/user/private-repo.git: exit status 128

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>

* regenerate proto files

Co-authored-by: Tonis Tiigi <tonistiigi@gmail.com>
Co-authored-by: Aaron Lehmann <alehmann@netflix.com>
Co-authored-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>
Co-authored-by: Pierre Fenoll <pierrefenoll@gmail.com>
Co-authored-by: Corey Larson <corey@earthly.dev>
Co-authored-by: Sebastiaan van Stijn <github@gone.nl>